bullets problem direction

+++ js/project/bullet.js
@@ -32,7 +32,6 @@ Bullet.prototype = {
 
         Bullet.count++;
         Bullet.bullets.push(this);
-        console.log('bk');
     },
 
     Render: function() {
@@ -123,23 +122,16 @@ Bullet.prototype = {
                 y3 = tank.yCoord,
                 y4 = tank.yCoord + Tank.sideSize;
 
-                //console.log('tanks: ' + Tank.tanks);
-                //console.log('tank id: ' + tank.id);
-                //console.log('b: ' + x1 + '_' + y1 + ':' +  x2 + '_' + y1  + ':' +  x2 + '_' + y2  + ':' + x1 + '_' + y2);
-                //console.log('t: ' + x3 + '_' + y3+ ':' +  x4 + '_' + y3  + ':' +  x4 + '_' + y4  + ':' + x3 + '_' + y4);
-                //alert('tank id: ' + tank.id);
-
-
-            if(tank.id != self.ownerId){
-                console.log(tank.id + '--' + self.id);                if(((y2 >= y3 && y2 <= y4) && ((x2 >= x3 && x2 <= x4) || (x1 <= x4 && x1 >= x3))) || 
+            if(tank.id != self.ownerId){              
+                if(((y2 >= y3 && y2 <= y4) && ((x2 >= x3 && x2 <= x4) || (x1 <= x4 && x1 >= x3))) || 
                   ((y4 >= y1 && y4 <= y2) && ((x4 >= x1 && x4 <= x2) || (x3 <= x2 && x3 >= x1))) ||
                   ((x2 >= x3 && x2 <= x4) && ((y2 >= y3 && y2 <= y4) || (y1 <= y4 && y1 >= y3))) ||
                   ((x4 >= x1 && x4 <= x2) && ((y4 >= y1 && y4 <= y2) || (y3 <= y2 && y3 >= y1)))) {
                         self.deleteObject(self);
                         self.deleteElement(self);
                         console.log('damage' + tank.id);
-                        //tank.deleteObject(tank);
-                        //tank.deleteElement(tank.tankId);
+                        tank.deleteObject(tank);
+                        tank.deleteElement(tank);
                         
                 };   
             };         
@@ -147,10 +139,10 @@ Bullet.prototype = {
     },                    
 
     Move: function() {
-      this.setBulletDirection();
-      this.offsetCalculate();
-      this.checkBorderCollision();
-      this.checkTankCollision();
-      this.Render();      
+        this.setBulletDirection();
+        this.offsetCalculate();
+        this.checkBorderCollision();
+        this.checkTankCollision();
+        this.Render();      
     }    
 }
